Create AutoInvoicer3.0.py
This is a Python script I built to streamline the process of renaming Invoice files at my current job. Since our documents are transferred to our computers via email, the files arrive with a default filename (such as 'doc543286794367834'), Our accountant was required to rename these files based on the Date and Invoice number. This script effectively automates this process. It requires a text file called 'company_names.txt' that is used as a list for all the company names that the invoices are for (which is used to name the directory that is created to store the files). This was one of my first ever python scripts, and I learned a lot about Cartesian coordinates and PDF file management in general